Fixed new screens and component bugs
Two new screens added: Create Game screen that is finished (for now) and
Lobby Screen that is just a dummy screen and needs components.
Fixed component bugs: InputFields now are clickable and can get focus
correctly. Also the marker is not there unless the field has focus
Buttons are not updated when not enabled, and the text is different when
they're not enabled.