Finished Working PieMenu
It works with hovering, updates correct clickedItemIndex wich is gotten
only when the menu is NOT visible (and the user is finished choosing or
not choosing an item). The last clickedItemIndex is gotten from the method
GetAndResetClickedItemIndex().
There are constants prefixed by PieMenu::C_ITEM_ for each item chosen, no
item chosen and the total amount of items