Lots and lots of small fixes, Credits screen
Added a new credits screen, with title, our names and a back button.
Credits button on menu screen now points there and back button goes back.

Changed order of Cancel and Start buttons in Join Game (now the cancel
button is to the right like on all the other screens).

Changes to CreateGame-, JoinGame- and Lobby states:
- Added title Labels
- Changed the size, alignment and positions of labels
- Changed positioning of all other components

Labels now optionally take the text size in the constructor and have a new
SetFont()-method.
InputFields now have an optional maximum number of characters and play a
sound when the user tries to write additional characters.
Replaced the click-sound and added a new inputFull-sound.
ClickMenu now can access items based on their index IN ADDITION to
accessing by Caption. The caption can also be changed.